changed revision from interface to generic class
